+++ packages/rma-server/src/print/aggregates/print-aggregate/print-aggregate.service.ts
@@ -94,21 +94,17 @@ export class PrintAggregateService {
       .text('Customer:', 50, customerInformationTop)
       .text(invoice.customer_name, 100, customerInformationTop, { width: 200 })
       .text('Address:', 50, customerInformationTop + 15)
-      .text(invoice.address, 100, customerInformationTop + 15)
-      .text('Contact:', 50, customerInformationTop + 30)
-      .text(invoice.contact, 100, customerInformationTop + 30)
+      .text(invoice.address_display, 100, customerInformationTop + 15)
       .text('Mobile No:', 50, customerInformationTop + 45)
-      .text(invoice.contact, 100, customerInformationTop + 45)
+      .text(invoice.contact_mobile, 100, customerInformationTop + 45)
 
       .text('Posting Date:', 300, customerInformationTop)
       .text(invoice.posting_date, 420, customerInformationTop)
-      .text('Due Date:', 300, customerInformationTop + 15)
-      .text(invoice.due_date, 420, customerInformationTop + 15)
       .text('Sold By', 300, customerInformationTop + 30)
-      .text(invoice.sold_by, 420, customerInformationTop + 30)
+      .text(invoice.sales_person, 420, customerInformationTop + 30)
       .text('Created By:', 300, customerInformationTop + 45)
       .text(invoice.created_by, 420, customerInformationTop + 45)
-      .text('Approved By:', 300, customerInformationTop + 60)
+      .text('Delivered By:', 300, customerInformationTop + 60)
       .text(invoice.modified_by, 420, customerInformationTop + 60);
 
     if (invoice?.print?.s_warehouse || invoice?.print?.t_warehouse) {
@@ -219,7 +215,12 @@ export class PrintAggregateService {
   async generateFooter(doc, settings: ServerSettings) {
     doc.moveDown();
     const image = await this.getCDNImage(settings.footerImageURL);
-    doc.image(image, 30, doc.y, { width: settings.footerWidth });
+    doc.image(image, 20, doc.page.height - 50, {
+      lineBreak: false,
+      width: settings.footerWidth,
+    });
+    // replaced above line with below for footer needed at bottom of page incase if it breaks in sales changed it.
+    // doc.image(image, 30, doc.y, { width: settings.footerWidth });
   }
 
   generateTableRow(