+++ packages/rma-server/src/warranty-claim/aggregates/warranty-claim-aggregate/warranty-claim-aggregate.service.ts
@@ -132,7 +132,7 @@ export class WarrantyClaimAggregateService extends AggregateRoot {
     clientHttpRequest,
   ) {
     return this.warrantyClaimsPoliciesService
-      .validateWarrantyCustomer(claimsPayload.customer)
+      .validateWarrantyCustomer(claimsPayload.customer_code)
       .pipe(
         switchMap(() => {
           return this.warrantyClaimsPoliciesService.validateWarrantySerailNo(
@@ -161,7 +161,7 @@ export class WarrantyClaimAggregateService extends AggregateRoot {
 
   createNonSerialClaim(claimsPayload: WarrantyClaimDto, clientHttpRequest) {
     return this.warrantyClaimsPoliciesService
-      .validateWarrantyCustomer(claimsPayload.customer)
+      .validateWarrantyCustomer(claimsPayload.customer_code)
       .pipe(
         switchMap(() => {
           return this.assignFields(claimsPayload, clientHttpRequest);
@@ -698,20 +698,37 @@ export class WarrantyClaimAggregateService extends AggregateRoot {
       }),
     );
   }
-  cancelWarrantyClaim(uuid: string) {
-    return this.warrantyClaimsPoliciesService.validateCancelClaim(uuid).pipe(
-      switchMap(res => {
-        return from(
-          this.warrantyClaimService.updateOne(
-            { uuid },
-            {
-              $set: {
-                claim_status: 'Cancelled',
+  cancelWarrantyClaim(cancelPayload: { uuid: string; serial_no: string }) {
+    return this.warrantyClaimsPoliciesService
+      .validateCancelClaim(cancelPayload.uuid)
+      .pipe(
+        switchMap(res => {
+          return from(
+            this.warrantyClaimService.updateOne(
+              { uuid: cancelPayload.uuid },
+              {
+                $set: {
+                  claim_status: 'Cancelled',
+                },
               },
-            },
-          ),
-        );
-      }),
-    );
+            ),
+          );
+        }),
+        switchMap(res => {
+          if (cancelPayload.serial_no) {
+            return from(
+              this.serialNoService.updateOne(
+                { serial_no: cancelPayload.serial_no },
+                {
+                  $unset: {
+                    claim_no: '',
+                  },
+                },
+              ),
+            );
+          }
+          return of({});
+        }),
+      );
   }
 }