+++ packages/rma-server/src/warranty-claim/command/reset-warranty-claim/reset-warranty-claim.command.ts
@@ -1,5 +1,5 @@
 import { ICommand } from '@nestjs/cqrs';
 
 export class ResetWarrantyClaimCommand implements ICommand {
-  constructor(public readonly uuid: string) {}
+  constructor(public readonly payload: { uuid: string; serial_no: string }) {}
 }