+++ packages/rma-warranty/src/app/warranty-ui/warranty/warranty.page.ts
@@ -23,6 +23,7 @@ import {
 } from '../../constants/app-string';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { CsvJsonService } from '../../api/csv-json/csv-json.service';
+import { APP_URL } from '../../constants/storage';
 
 @Component({
   selector: 'app-warranty',
@@ -48,7 +49,6 @@ export class WarrantyPage implements OnInit {
     'claim_no',
     'claim_type',
     'received_date',
-    'customer_third_party',
     'customer_name',
     'third_party_name',
     'item_code',
@@ -59,7 +59,6 @@ export class WarrantyPage implements OnInit {
     'received_by',
     'delivered_by',
     'product_brand',
-    'damaged_serial',
     'replace_serial',
     'problem',
     'verdict',
@@ -311,7 +310,18 @@ export class WarrantyPage implements OnInit {
         );
         break;
       default:
-        this.router.navigate(['/warranty/view-warranty-claims', row.uuid]);
+        this.warrantyService
+          .getStorage()
+          .getItem(APP_URL)
+          .then(url => {
+            this.router.navigate([]).then(result => {
+              window.open(
+                `${url}/warranty/view-warranty-claims/${row.uuid}`,
+                '_blank',
+              );
+            });
+          });
+
         break;
     }
   }